requirements/dev.txt
-r base.txt
ipython==8.29.0
django-extensions==3.2.3